Add global config file with api_conn_timeout param. Closes #4
Add a global settings.json in ~/.henry/ that allows users to define
settings globally. Currently only the "api_conn_timeout" setting is
recognised and it can take 3 types of values: null, an integer representing
connect and read timeouts (in seconds) combined or a list that specifies
the connect and read timeouts separately. The settings.json file should look
as follows:
{
  "api_conn_timeout": x
}
where x can be null, int or [int, int].